blog/requirements.txt
mysql-connector-python==8.0.13
Faker==1.0.0
pyjwt==1.6.4